\name{r_sizes}
\alias{r_sizes}
\docType{data}
\title{
r_sizes
}
\description{
Numeric variable containing the sizes of the reference peaks (ABI GS500LIZ). 
Peaks not to be counted are labeled \code{NA}.
}
\usage{data(r_sizes)}
\format{
  The format is:
 num [1:15] 50 75 100 139 150 160 200 NA 300 NA ...
}
\details{
}
\source{
}
\references{
\code{citation(package="seqinr")}

Applied Biosystems. (2009) Applied Biosystems Genetic Analysis Data File Format [Online].  Available from: http://www.appliedbiosystems.com/support/software_community/ABIF_File_Format.pdf (Accessed: 14 January 2010). 
}
\seealso{
\code{searchStandardPeaks}, \code{sizeCaller}
}
\examples{
r_sizes
}
