\name{LazyMLVA-package}
\alias{LazyMLVA-package}
\alias{LazyMLVA}
\docType{package}
\title{
LazyMLVA
}
\description{
The package automates assignment of MLVA types. MLVA stands for 'Multiple-Locus-VNTR-Analysis', which is a typing approach for bacterial pathogens. It extracts the size standard from ABIF electropherogram files and automatically looks for products of unknown sizes in the specified channels. Further to estimating the sizes of found products, it converts the sizes to repeat lengths according to data specified in a 'lazy map'. The package depends on seqinr.
}
\details{
\tabular{ll}{
Package: \tab LazyMLVA\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2010-01-14\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Johannes Elias

Maintainer: Johannes Elias <jelias@hygiene.uni-wuerzburg.de>
}
\references{
\code{citation(package="seqinr")}

Applied Biosystems. (2009) Applied Biosystems Genetic Analysis Data File Format [Online].  Available from: http://www.appliedbiosystems.com/support/software_community/ABIF_File_Format.pdf (Accessed: 14 January 2010). 

}
\keyword{ package }
\seealso{
}
\examples{
}
